Move step logic to the step package
