--- additional_options.go
@@ -1,21 +1,0 @@
-package main
-
-import "github.com/bitrise-io/go-utils/sliceutil"
-
-func generateAdditionalOptions(platform string, customOptions []string) []string {
-	destination := "generic/platform=" + platform
-	destinationOptions := []string{"-destination", destination}
-
-	var options []string
-	if len(customOptions) != 0 {
-		if !sliceutil.IsStringInSlice("-destination", customOptions) {
-			options = append(options, destinationOptions...)
-		}
-
-		options = append(options, customOptions...)
-	} else {
-		options = append(options, destinationOptions...)
-	}
-
-	return options
-}
